+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## unreleased
 - added the ability to run flutter apps in debug / profile / release modes
 - added the ability to connect to already running Flutter apps on devices (currently Android only)
+- split the mojo launch config out from the flutter one
 
 ## 0.6.24
 - fix an issue with newly created launches